1. Make sure directories  of `logs` and `pods` exist before logging. 2. Refactor: put `startLogging` into Common.
