+++ lib/Common.js
@@ -8,6 +8,7 @@ var util      = require('util');
 var cronJob   = require('cron').CronJob;
 var isBinary  = require('isbinaryfile');
 
+var async     = require('async');
 var UX        = require('./CliUx.js');
 var cst       = require('../constants.js');
 var extItps   = require('./interpreter.json');
@@ -22,6 +23,61 @@ var InteractorDaemonizer = require('./Interactor/InteractorDaemonizer.js');
  */
 
 var Common = module.exports;
+
+/**
+ * Start log outgoing messages
+ * @method startLogging
+ * @param {} callback
+ * @return
+ */
+Common.startLogging = function(stds, callback) {
+  // Make sure directories of `logs` and `pids` exist.
+  try {
+    ['logs', 'pids'].forEach(function(n){
+      (function(_path){
+        !fs.existsSync(_path) && fs.mkdirSync(_path, '0755');
+      })(path.resolve(cst.PM2_ROOT_PATH, n))
+    });
+  }catch(err){
+    return callback(new Error('can not create directories (logs/pids):' + err.message));
+  }
+
+  // waterfall.
+  var flows = [];
+  // types of stdio, should be sorted as `std(entire log)`, `out`, `err`.
+  var types = Object.keys(stds).sort(function(x, y){
+    return -x.charCodeAt(0) + y.charCodeAt(0);
+  });
+
+  // Create write streams.
+  (function createWS(io){
+    if(io.length != 1){
+      return;
+    }
+    io = io[0];
+
+    // If `std` is a Stream type, try next `std`.
+    // compatible with `pm2 reloadLogs`
+    if(typeof stds[io] == 'object' && !isNaN(stds[io].fd)){
+      return createWS(types.splice(0, 1));
+    }
+
+    flows.push(function(next){
+      var file = stds[io];
+      stds[io] = fs.createWriteStream(file, {flags: 'a'})
+        .on('error', function(err){
+          next(err);
+        })
+        .on('open', function(){
+          next();
+        });
+      stds[io]._file = file;
+    });
+    createWS(types.splice(0, 1));
+  })(types.splice(0, 1));
+
+  async.waterfall(flows, callback);
+}
 
 /**
  * Resolve app paths and replace missing values with defaults.
