+++ lib/God/ForkMode.js
@@ -8,7 +8,6 @@
 
 var log           = require('debug')('pm2:god');
 var fs            = require('fs');
-var async         = require('async');
 var cst           = require('../../constants.js');
 var moment        = require('moment');
 var Common        = require('../Common');
@@ -74,56 +73,7 @@ module.exports = function ForkMode(God) {
       stds.std = pm2_env.pm_log_path;
     }
 
-    /**
-     * Description
-     * @method startLogging
-     * @param {} cb
-     * @return
-     */
-    function startLogging(cb) {
-      // waterfall.
-      var flows = [];
-      // types of stdio, should be sorted as `std(entire log)`, `out`, `err`.
-      var types = Object.keys(stds).sort(function(x, y){
-        return -x.charCodeAt(0) + y.charCodeAt(0);
-      });
-
-      // Create write streams.
-      (function createWS(io){
-        if(io.length != 1){
-          return;
-        }
-        io = io[0];
-
-        // If `std` is a Stream type, try next `std`.
-        // compatible with `pm2 reloadLogs`
-        if(typeof stds[io] == 'object' && !isNaN(stds[io].fd)){
-          return createWS(types.splice(0, 1));
-        }
-
-        flows.push(function(next){
-          var file = stds[io];
-          stds[io] = fs.createWriteStream(file, {flags: 'a'})
-            .on('error', function(e){
-              next(e);
-            })
-            .on('open', function(){
-              next();
-            });
-          stds[io]._file = file;
-        });
-        createWS(types.splice(0, 1));
-      })(types.splice(0, 1));
-
-      async.waterfall(flows, function(err, result){
-        if (err){
-          God.logAndGenerateError(err);
-        }
-        cb(err);
-      });
-    }
-
-    startLogging(function(err) {
+    Common.startLogging(stds, function(err, result) {
       if (err) {
         God.logAndGenerateError(err);
         return cb(err);
@@ -219,7 +169,7 @@ module.exports = function ForkMode(God) {
           stds[k] = stds[k]._file;
         }
         cspr.removeAllListeners();
-        startLogging(cb);
+        Common.startLogging(stds, cb);
       };
 
       cspr.unref();
