+++ lib/ProcessContainer.js
@@ -8,8 +8,8 @@ if (process.env.name != null)
 
 var fs     = require('fs');
 var p      = require('path');
-var async  = require('async');
 var cst    = require('../constants');
+var Common = require('./Common');
 var axm    = require('axm');
 
 /**
@@ -88,8 +88,12 @@ function exec(script, stds) {
         stds[k].close();
         stds[k] = stds[k]._file;
       }
-      startLogging(function () {
-        console.log('Reloading log...');
+      Common.startLogging(stds, function (err) {
+        if(err){
+          console.error('Failed to reload logs:', err.stack);
+        }else {
+          console.log('Reloading log...');
+        }
       });
     }
   });
@@ -99,51 +103,7 @@ function exec(script, stds) {
   if (process.env.log_date_format)
     moment = require('moment');
 
-
-  /**
-   * Description
-   * @method startLogging
-   * @param {} callback
-   * @return
-   */
-  function startLogging(callback) {
-    // waterfall.
-    var flows = [];
-    // types of stdio, should be sorted as `std(entire log)`, `out`, `err`.
-    var types = Object.keys(stds).sort(function(x, y){
-      return -x.charCodeAt(0) + y.charCodeAt(0);
-    });
-
-    // Create write streams.
-    (function createWS(io){
-      if(io.length != 1){
-        return;
-      }
-      io = io[0];
-
-      // If `std` is a Stream type, try next `std`.
-      // compatible with `pm2 reloadLogs`
-      if(typeof stds[io] == 'object' && !isNaN(stds[io].fd)){
-        return createWS(types.splice(0, 1));
-      }
-
-      flows.push(function(next){
-        var file = stds[io];
-        stds[io] = fs.createWriteStream(file, {flags: 'a'})
-          .on('error', function(e){
-            next(e);
-          }).on('open', function(){
-            next();
-          });
-        stds[io]._file = file;
-      });
-      createWS(types.splice(0, 1));
-    })(types.splice(0, 1));
-
-    async.waterfall(flows, callback);
-  }
-
-  startLogging(function (err) {
+  Common.startLogging(stds, function (err) {
     if (err) {
       process.send({
         type    : 'process:exception',
